requirements.txt
langchain
langgraph
langchain-openai
langchain-community
python-dotenv
PyMuPDF
tqdm
sentence-transformers
scikit-learn
openai
streamlit
unstructured[local-inference,pdf]
